+++ yash-builtin/CHANGELOG.md
@@ -13,6 +13,9 @@ A _private dependency_ is used internally and not visible to downstream users.
 
 ### Changed
 
+- The `command` built-in now requires a `yash_env::semantics::command::RunFunction`
+  instance to be available in the environment's `any` storage. This instance is
+  used to invoke shell functions in the `command::Invoke::execute` method.
 - The `eval` built-in now requires a `yash_env::semantics::RunReadEvalLoop`
   instance to be available in the environment's `any` storage. This instance is
   used to run the read-eval loop in the `eval::main` function.